remove NpowAddressMapping trait
since `reward_mapping` now in pallet-deeper-node, it's better to also
move `get_npow_reward` to pallet-deeper-node
pallet-operation doesn't depends on pallet-evm
