+++ pallets/staking/src/mock.rs
@@ -282,7 +282,6 @@ impl pallet_operation::Config for Test {
     type OPWeightInfo = ();
     type MinimumBurnedDPR = MinimumBurnedDPR;
     type CreditInterface = Credit;
-    type NpowAddressMapping = ();
     type UserPrivilegeInterface = UserPrivileges;
 }
